Improve granularity of THCCachingHostAllocator locking.
In cases where copyAsync is a large percentage of the work,
processing events in recordEvent can cause a large bottleneck.

Here, instead of a single lock, we use two main locks:
1) lock around used/available blocks (main mutex)
2) lock around outstanding cuda_events

We also relax the aggressiveness of looking for free blocks
in recordEvent (copyAsync), by ensuring that one thread is
processing events instead of having each thread wait each time.